- some FreeRDP issues already fixed with myrtille (printer, audio) are now also fixed into the FreeRDP master branch. Some are still yet to fix (gdi, per device license mode, FQDN resolve with NLA, etc.).  TODO: pull requests!
- removed attach console to parent process into wfreerdp
- refactored clipboard handling with myrtille

+++ channels/rdpgfx/client/rdpgfx_main.h
@@ -72,7 +72,7 @@ struct _RDPGFX_PLUGIN
 	ZGFX_CONTEXT* zgfx;
 	UINT32 UnacknowledgedFrames;
 	UINT32 TotalDecodedFrames;
-	UINT32 StartDecodingTime;
+	UINT64 StartDecodingTime;
 	BOOL suspendFrameAcks;
 	BOOL sendFrameAcks;
 
@@ -89,4 +89,3 @@ struct _RDPGFX_PLUGIN
 typedef struct _RDPGFX_PLUGIN RDPGFX_PLUGIN;
 
 #endif /* FREERDP_CHANNEL_RDPGFX_CLIENT_MAIN_H */
-
